Added support for B2C authorities per #195
