+++ tests/Test.MSAL.NET.Unit/InstanceTests/AuthorityTests.cs
@@ -71,7 +71,7 @@ namespace Test.MSAL.NET.Unit.InstanceTests
             HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
             {
                 Method = HttpMethod.Get,
-                Url = "https://login.windows.net/common/discovery/instance",
+                Url = "https://login.microsoftonline.com/common/discovery/instance",
                 QueryParams = new Dictionary<string, string>
                 {
                     {"api-version", "1.0"},
@@ -143,7 +143,7 @@ namespace Test.MSAL.NET.Unit.InstanceTests
             HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
             {
                 Method = HttpMethod.Get,
-                Url = "https://login.windows.net/common/discovery/instance",
+                Url = "https://login.microsoftonline.com/common/discovery/instance",
                 QueryParams = new Dictionary<string, string>
                 {
                     {"api-version", "1.0"},
@@ -240,338 +240,6 @@ namespace Test.MSAL.NET.Unit.InstanceTests
                 Task.Run(async () =>
                 {
                     await instance.ResolveEndpointsAsync(null, new RequestContext(Guid.NewGuid()));
-                }).GetAwaiter().GetResult();
-                Assert.Fail("validation should have failed here");
-            }
-            catch (MsalServiceException exc)
-            {
-                Assert.AreEqual(MsalError.TenantDiscoveryFailed, exc.ErrorCode);
-            }
-
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsSuccessfulValidationUsingOnPremiseDrsTest()
-        {
-            //add mock response for on-premise DRS request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://enterpriseregistration.fabrikam.com/enrollmentserver/contract",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"api-version", "1.0"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("drs-response.json"))
-            });
-
-
-            //add mock response for on-premise webfinger request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.fabrikam.com/adfs/.well-known/webfinger",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"resource", "https://fs.contoso.com"},
-                    {"rel", "http://schemas.microsoft.com/rel/trusted-realm"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessWebFingerResponseMessage()
-            });
-
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.contoso.com/adfs/.well-known/openid-configuration",
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("OpenidConfiguration-OnPremise.json"))
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, true);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            Task.Run(async () =>
-            {
-                await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-            }).GetAwaiter().GetResult();
-
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/authorize/",
-                instance.AuthorizationEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/token/",
-                instance.TokenEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs",
-                instance.SelfSignedJwtAudience);
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-            Assert.AreEqual(1, Authority.ValidatedAuthorities.Count);
-
-            //attempt to do authority validation again. NO network call should be made
-            instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, true);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            Task.Run(async () =>
-            {
-                await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-            }).GetAwaiter().GetResult();
-
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/authorize/",
-                instance.AuthorizationEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/token/",
-                instance.TokenEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs",
-                instance.SelfSignedJwtAudience);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsSuccessfulValidationUsingCloudDrsFallbackTest()
-        {
-            //add mock failure response for on-premise DRS request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://enterpriseregistration.fabrikam.com/enrollmentserver/contract",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"api-version", "1.0"}
-                },
-                ResponseMessage = MockHelpers.CreateFailureMessage(HttpStatusCode.NotFound, "not found")
-            });
-
-            //add mock response for cloud DRS request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://enterpriseregistration.windows.net/fabrikam.com/enrollmentserver/contract",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"api-version", "1.0"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("drs-response.json"))
-            });
-
-
-            //add mock response for on-premise webfinger request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.fabrikam.com/adfs/.well-known/webfinger",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"resource", "https://fs.contoso.com"},
-                    {"rel", "http://schemas.microsoft.com/rel/trusted-realm"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessWebFingerResponseMessage()
-            });
-
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.contoso.com/adfs/.well-known/openid-configuration",
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("OpenidConfiguration-OnPremise.json"))
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, true);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            Task.Run(async () =>
-            {
-                await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-            }).GetAwaiter().GetResult();
-
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/authorize/",
-                instance.AuthorizationEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/token/",
-                instance.TokenEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs",
-                instance.SelfSignedJwtAudience);
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsValidationOffSuccessTest()
-        {
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.contoso.com/adfs/.well-known/openid-configuration",
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("OpenidConfiguration-OnPremise.json"))
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, false);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            Task.Run(async () =>
-            {
-                await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-            }).GetAwaiter().GetResult();
-
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/authorize/",
-                instance.AuthorizationEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs/oauth2/token/",
-                instance.TokenEndpoint);
-            Assert.AreEqual("https://fs.contoso.com/adfs",
-                instance.SelfSignedJwtAudience);
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsFailedValidationTest()
-        {
-            //add mock response for on-premise DRS request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://enterpriseregistration.fabrikam.com/enrollmentserver/contract",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"api-version", "1.0"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("drs-response.json"))
-            });
-
-
-            //add mock response for on-premise webfinger request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.fabrikam.com/adfs/.well-known/webfinger",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"resource", "https://fs.contoso.com"},
-                    {"rel", "http://schemas.microsoft.com/rel/trusted-realm"}
-                },
-                ResponseMessage = MockHelpers.CreateFailureMessage(HttpStatusCode.NotFound, "not-found")
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, true);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            try
-            {
-                Task.Run(async () =>
-                {
-                    await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-                }).GetAwaiter().GetResult();
-                Assert.Fail("ResolveEndpointsAsync should have failed here");
-            }
-            catch (Exception exc)
-            {
-                Assert.IsNotNull(exc);
-            }
-
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsFailedValidationResourceNotInTrustedRealmTest()
-        {
-            //add mock response for on-premise DRS request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://enterpriseregistration.fabrikam.com/enrollmentserver/contract",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"api-version", "1.0"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("drs-response.json"))
-            });
-
-
-            //add mock response for on-premise webfinger request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.fabrikam.com/adfs/.well-known/webfinger",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"resource", "https://fs.contoso.com"},
-                    {"rel", "http://schemas.microsoft.com/rel/trusted-realm"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessWebFingerResponseMessage("https://fs.some-other-sts.com")
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, true);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            try
-            {
-                Task.Run(async () =>
-                {
-                    await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-                }).GetAwaiter().GetResult();
-                Assert.Fail("ResolveEndpointsAsync should have failed here");
-            }
-            catch (Exception exc)
-            {
-                Assert.IsNotNull(exc);
-            }
-
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsFailedValidationMissingFieldsInDrsResponseTest()
-        {
-            //add mock failure response for on-premise DRS request
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://enterpriseregistration.fabrikam.com/enrollmentserver/contract",
-                QueryParams = new Dictionary<string, string>
-                {
-                    {"api-version", "1.0"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("drs-response-missing-field.json"))
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, true);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            try
-            {
-                Task.Run(async () =>
-                {
-                    await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
-                }).GetAwaiter().GetResult();
-                Assert.Fail("ResolveEndpointsAsync should have failed here");
-            }
-            catch (Exception exc)
-            {
-                Assert.IsNotNull(exc);
-            }
-
-            Assert.AreEqual(0, HttpMessageHandlerFactory.MockCount);
-        }
-
-        [TestMethod]
-        [TestCategory("AuthorityTests")]
-        public void AdfsFailedTenantDiscoveryMissingEndpointsTest()
-        {
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                Url = "https://fs.contoso.com/adfs/.well-known/openid-configuration",
-                ResponseMessage = MockHelpers.CreateSuccessResponseMessage(File.ReadAllText("OpenidConfiguration-MissingFields-OnPremise.json"))
-            });
-
-            Authority instance = Authority.CreateAuthority(TestConstants.OnPremiseAuthority, false);
-            Assert.IsNotNull(instance);
-            Assert.AreEqual(instance.AuthorityType, AuthorityType.Adfs);
-            try
-            {
-                Task.Run(async () =>
-                {
-                    await instance.ResolveEndpointsAsync(TestConstants.FabrikamDisplayableId, new RequestContext(Guid.NewGuid()));
                 }).GetAwaiter().GetResult();
                 Assert.Fail("validation should have failed here");
             }
